feat: change user.image default value
Modify the logic of getting default profile image of an user.

The default profile image was retrieved from the default value set
in CustomUser.image field. This has the drawback that the default
profile image lies in 'media/' folder, which should not be tracked
by git because it varies in different developer's setting. Instead,
the default profile image now is in 'static' folder, and the field
CustomUser.image removes a default value.